#54 [feat] 프로필 입력화면 리팩토링
