+++ app/src/main/java/com/spark/android/data/repository/ProfileRepository.kt
@@ -1,0 +1,4 @@
+package com.spark.android.data.repository
+
+interface ProfileRepository {
+}
