+++ app/src/main/java/com/spark/android/ui/profile/ProfileFragment.kt
@@ -14,7 +14,9 @@ import com.spark.android.util.KeyBoardUtil
 import com.spark.android.util.initStatusBarColor
 import com.spark.android.util.initStatusBarTextColorToWhite
 import com.spark.android.util.popBackStack
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class ProfileFragment : BaseFragment<FragmentProfileBinding>(R.layout.fragment_profile) {
     private val profileViewModel by viewModels<ProfileViewModel>()
